reward for the current action not the previous
mario was getting a reward for the previous action when performing the current action. This means that walking into enemy was given the positive reward of moving forward(the action before moving into the enemy), where as the action after hitting the enemy(moving forward) was given a huge negative reward.
The existence of this problem was first thought about on my journey on the luas this morning, and was later found to be true as I read through the code this afternoon in LG12.